chore: refactor login page logic
